improve warning system
