+++ python/taichi/misc/gui.py
@@ -1,6 +1,7 @@
 import numbers
 import numpy as np
 from taichi import ti_core
+from .util import deprecated
 
 
 class GUI:
@@ -285,12 +286,8 @@ class GUI:
         pos = self.core.get_cursor_pos()
         return pos[0], pos[1]
 
+    @deprecated('gui.has_key_pressed()', 'gui.get_event()')
     def has_key_pressed(self):
-        import warnings
-        warnings.warn(
-            'gui.has_key_pressed() is deprecated, use gui.get_event() instead.',
-            DeprecationWarning,
-            stacklevel=3)
         if self.has_key_event():
             self.get_key_event()  # pop to update self.key_pressed
         return len(self.key_pressed) != 0
